// FrontEnd/amaris-app/src/app/services/fondos-api/model/backendFondosApplicationDTOsUsuarioDto.ts
/**
 * BackendFondos
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BackendFondosApplicationDTOsUsuarioDto { 
    usuarioID?: string;
    nombreUsuario?: string;
    email?: string;
    rol?: string;
    password?: string;
}